Add upstream

+++ examples/src/rescue_raps/custom_trace_table.rs
@@ -1,4 +1,5 @@
 // Copyright (c) Facebook, Inc. and its affiliates.
+// Copyright (c) 2021-2023 Toposware, Inc.
 //
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
@@ -85,7 +86,7 @@ impl<B: StarkField> RapTraceTable<B> {
             "execution trace length must be a power of 2"
         );
         assert!(
-            log2(length) as u32 <= B::TWO_ADICITY,
+            log2(length) <= B::TWO_ADICITY,
             "execution trace length cannot exceed 2^{} steps, but was 2^{}",
             B::TWO_ADICITY,
             log2(length)
